DBTSetup component
